media: s5p-mfc: use bidirectional DMA flag for decoding to dmabuf
In case we are requested to write the decoded data into a buffer
provided by dmabuf, we need to map such buffer bidirectional.
The MFC hardware stores temporary decoding data inside the unused
parts of the buffer, hence not only reading from it, but also
writing.

In particular, if not mapped bidirectional, this triggers an IOMMU
pagefault as soon as we enable streaming on the OUTPUT queue.